fixing template filters
